feat: Add programs for printing a character, character input/output, and new line
This commit adds three new programs to the repository:
- Program 01 - Print A Character
- Program 02 - Character Input and Output
- Program 04 - New Line

These programs demonstrate basic input/output operations in assembly language. They showcase how to display a character, get input from the user, and move to a new line. The programs are well-documented and can be used as a reference for learning assembly language programming.